hyperactor_telemetry::env

Function execution_id

pub fn execution_id() -> String

Forward or generate a uuid for this execution. When running in production on mast, this is provided to us via the MAST_HPC_JOB_NAME env var. Subprocesses should either forward the MAST_HPC_JOB_NAME variable, or set the “MONARCH_EXECUTION_ID” var for subprocesses launched by this process. We keep these env vars separate so that other applications that depend on the MAST_HPC_JOB_NAME existing to understand their environment do not get confused and think they are running on mast when we are doing local testing.