hyperactor_telemetry

Function initialize_logging

pub fn initialize_logging(clock: impl TelemetryClock + Send + 'static)

Set up logging based on the given execution environment. We specialize logging based on how the logs are consumed. The destination scuba table is specialized based on the execution environment. mast -> monarch_tracing/prod devserver -> monarch_tracing/local unit test -> monarch_tracing/test scuba logging won’t normally be enabled for a unit test unless we are specifically testing logging, so you don’t need to worry about your tests being flakey due to scuba logging. You have to manually call initialize_logging() to get this behavior.